ReturnArray = Range("A2:L7").Value; Prac.bas
It is more efficient to First add the cell values of a Worksheet into an array, then manipulate the contents of the array. Commit also includes several practice routines by means of identifying the last row with a value and column with a value inclusive or exclusive of blank cells. Additionally, reviewed the VBA Library's Information Class members within the Object Browser and preformed test evaluations.